refactor: AttendanceTable을 공용 컴포넌트로 변경 #1
